the paper section is done.
This is a group effort with Joseph Barbosa. He has handled the web end of our design and the design of the device and I have written the paper portion of the project.

On my repository you will find the the .tex file and pdf for the paper as well as image source files and on Joseph's repository you will find the web page with design information as well as a link to the PDF version of the paper.